+++ CHANGELOG.md
@@ -7,7 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [2.6.5] - 2024-03-011
+## [2.7.0] - 2024-04-20
+### Added
+- Added the following new arguments:
+  - `--rundeck.projects.nodes.info` - If passed, display Rundeck projects nodes info metrics, currently only the `rundeck_project_nodes_total` metric is available.
+    - Requests for this check are cached as it can cause high CPU load depending on the number of projects.
+  - `--threadpool_max_workers` - The maximum number of workers in the threadpool to run rundeck_exporter asynchronous checks.
+    - Defaults to `(number of CPUs) + 4`, which may be too much on a server running other services.
+  - `--rundeck.requests.timeout` - The maximum number of seconds that requests to the Rundeck API should timeout.
+    - Defaults to 30.
+
+### Changed
+- Changed the functions `request_data_from` and `cached_request_data_from` to `request` and `cached_request`, respectively.
+
+## [2.6.5] - 2024-03-11
 ### Added
 - Issue [#85](https://github.com/phsmith/rundeck_exporter/issues/85), added new metric `rundeck_project_nodes_total`.
 
@@ -262,7 +275,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[unreleased]: https://github.com/phsmith/rundeck_exporter/compare/v2.6.5...HEAD
+[unreleased]: https://github.com/phsmith/rundeck_exporter/compare/v2.7.0...HEAD
+[2.7.0]: https://github.com/phsmith/rundeck_exporter/compare/v2.6.5...v2.7.0
 [2.6.5]: https://github.com/phsmith/rundeck_exporter/compare/v2.6.4...v2.6.5
 [2.6.4]: https://github.com/phsmith/rundeck_exporter/compare/v2.6.3...v2.6.4
 [2.6.3]: https://github.com/phsmith/rundeck_exporter/compare/v2.6.2...v2.6.3